TABDEV-1005. Don't use self or description as an identifier.
'self' is an Objective-C keyword, and while it is legal to use as a struct field, it is confusing.
'description' is an NSObject top-level method, and it seems unsafe since most code won't expect a setDescription: method to exist.

Changed 'self' to 'selfAttribute'. This should have no effect since all instances were private.
Changed 'description' to 'jiveDescription' in the same way that 'id' is changed to 'jiveId'