TABDEV-1568. Fixed broken build when archiving.
When archiving a static library target, Xcode really needs to be able to find a static library. The archive step blindly checks for one, whether the static library target built one or not. If the library doesn't exist, the build fails when trying to copy and chown the missing library. The fix is to build a dummy file with a dummy global symbol. I derived both from the target name.